fix(metadata): use metadata.getOrCreateOwn 's optional targettKey for TS to circumvent TS's 'extends' implementation
Typescript's 'extends' is on purpuse not according to specifications. Therefore, Reflects usage with targettKey===undefinded of hasOwnProperty(targettKey)  returns true for a parent property 'undefined' (The derived class doesn have that at that point, thus false should have been returned).

A simple fix for us is not to use the key 'undefined' (as per specification) in this case, but use a key on basis of the class name. Since they are different, the faulty 'extends' doesn't propose an issue