refactor(association-select): inject Container instead of Entitymanager
injecting internal dependencies can lead to problems with the load order in aurelia-cli resp. jspm. in such a case, the EntitiyManager class isn't defined at the time of @inject evaluation resulting in a DI error.. the only solution i found which worked for both, was to inject the container instead.